# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Code Challenge | 2DCH7-20059",
        "title": "Pars Tasmim",
        "contact": {
            "name": "Erfan Derakhshani",
            "url": "https://something.com",
            "email": "techerfan@gmail.com"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/users/login": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "login",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UserLoginResponse"
                        }
                    },
                    "400": {
                        "description": "bad request"
                    },
                    "401": {
                        "description": "unauthorized"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal error"
                    }
                }
            }
        },
        "/users/register": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Register a new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Register a new user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UserRegisterResponse"
                        }
                    },
                    "400": {
                        "description": "bad request"
                    },
                    "401": {
                        "description": "unauthorized"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal error"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.UserLoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
...
        "dto.UserRegisterResponse": {
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}